Simulator

+++ Engine/src/main/java/ru/nsu/team/entity/statistics/CarState.java
@@ -2,21 +2,21 @@ package ru.nsu.team.entity.statistics;
 
 import ru.nsu.team.entity.roadmap.Road;
 import ru.nsu.team.entity.trafficparticipant.Car;
-import ru.nsu.team.entity.trafficparticipant.Position;
+import ru.nsu.team.entity.roadmap.Position;
 import ru.nsu.team.entity.trafficparticipant.PositionOnRoad;
 import ru.nsu.team.entity.trafficparticipant.TrafficParticipant;
 
 public class CarState {
-  private int speed;
+  private double speed;
   private final int id;
   private double position;
-  private int time;
+  private long time;
   private int currentLane;
   private int currentRoad;
   private Position from;
   private Position to;
 
-  public CarState(TrafficParticipant trafficParticipant, int time) {
+  public CarState(TrafficParticipant trafficParticipant, long time) {
     Car car = trafficParticipant.getCar();
     PositionOnRoad positionOnRoad = trafficParticipant.getPosition();
     speed = car.getSpeed();
@@ -30,7 +30,7 @@ public class CarState {
     to = road.getTo().getPosition();
   }
 
-  public int getSpeed() {
+  public double getSpeed() {
     return speed;
   }
 
@@ -42,7 +42,7 @@ public class CarState {
     return position;
   }
 
-  public int getTime() {
+  public long getTime() {
     return time;
   }
